move highlight module into core
shouldn't be in utils as it uses the plugin config directly
